Remove maintenance component
